[JIRA SONIC-340] show platform syseeprom is not working
It has been confirmed that race condition does exist between
as7326-platform-handle_mac.service and opennsl-modules.service
as a result, upon 1st boot attempt after NOS installation, MAC and PCI reset
will not be performed and I2C virtual busses might not be created properly.

And thus, in this commit, we'll propose a platform_init.sh in /etc/rc.local,
and it will only be executed if it's present.

Signed-off-by: Dante (Kuo-Jung) Su <dante.su@broadcom.com>
Change-Id: I233dde60d3132a98a001343a31dd39206b3367d3