[build] Adjust SHA-1 calc for Docker user images
Each Docker user image is uniquely tagged based on the SHA-1 hash of
several items, including the user name, current working directory,
and the platform being built. The PLATFORM parameter is only present
when specified by 'make configure', but is otherwise undefined. Since
the intent is to use the platform as a differentiator, substitute the
current value of the .platform file (as created by make configure)
when the PLATFORM variable is unspecified.

Change-Id: I15a4c96c2e0319a37afb70eb51226245cbc8c971