[JIRA SONIC-835] SONIC_2.0 - seg fault after 'show ip interface' right after the vlan range add cmd
Added latest version of netifaces python library 0.10.9 to source tree.

Added a patch fix segfault while accessing netifaces library when large number of interfaces are in the process to be created.

Included this patched version of netifaces in base debian distro as well as the orchagent docker. Once this
patch gets upstreamed to netifaces repo, these changes will be rolled back and
pip install netifaces==<new version> will be used.

Signed-off-by: Rajendra Dendukuri <rajendra.dendukuri@broadcom.com>
Change-Id: Ic2114882f42e7044a8235a5fe21ede5e784c705d