[build] Enhance system locking to support native Docker daemon on shared server
A series of docker image saves and loads occurs during the build using generic
image names and the :latest tag. On a shared server, multiple builds may be
working with the same image name, requiring mutex protection that encompasses
both the saving of an image as well as the loading of a base image used
in the creation of another image.

Change-Id: I8e13d966a22e0df54cd5823b02b900e7b48a9c85