[JIRA SONIC-5184] ST:After clear ARP/ND, ARP/ND is not resolved if static route's next-hop is router port(Ethernet/Port channel as IP intf)
Root Cause:
Currently the arp_update script will refresh the neighbor entries in the
following cases:
  1. If there is NEIGH ENTRY in the config_db.
  2. For all VEs.
In this case, since the arp/ndp was resolved, the entry was removed
from the config_db and the L3 interface is a router port.
So, arp/ndp, once it moves to Failed state, is never refreshed.

UT Result:
Verified arp/ndp refresh on Pushaw-2/MH-2
Dev sanity
Sanity_Logs @
http://10.59.132.240:9009/projects/csg_sonic/sanity/logs/dev-sanity-td3-td3-02/sonic_sanity_190621_0241_745

Change-Id: I33f411a81685c0992be2f811085370569d59777b