[build] Rework original fix avoiding base image :latest tag to prepare it for upstreaming
The original fix to avoid the use of the :latest tag when searching for a base image
in a shared build server was hard-coded to get things working. In order to upstream
this fix, it has been reworked so that it's driven by a new DOCKER_AVOID_BASE_TAG_LATEST
option in rules/config. If the community likes this idea, these changes should make
it more acceptable for merging into the upstream.

Change-Id: Idca017d6ae853efef05c9b651e5f57573cf81aaa