[build] Build fsroot in two phases so first phase can occur earlier
This change admittedly pushes the limits of acceptability to save
a couple minutes of build time. This is not intended for upstreaming,
though the decision not to consolidate the phases in build_debian.sh
is to allow easier merges from upstream into the development branches.

The phase 1 items must not rely on any of the build artifacts, but
be restricted to fixed items such as installing Debian packages,
Docker, etc. into the fsroot.

Change-Id: Idb3a91d906198fcf42cfdf44011c735b075f672e