[JIRA SONIC-6498] Revise base image latest tag avoidance fix
This change is another modification to the fix that avoids using
the :latest tag when identifying the Docker base image to use when
creating a user image. Based on upstream PR feedback, the SONiC
maintainers want this change to always be in effect, rather than
enabled through a build option. Therefore, the rules/config
DOCKER_AVOID_BASE_TAG_LATEST option has been removed.

Since this solution requires the "ARG before FROM" support in
Docker, a minimum version check is added in Makefile.work to
ensure the build host is running with an adequate Docker
server version.

Change-Id: I944de2ade2490b9408f10bb829f4eae84de44e9f