[JIRA SONIC-6503] SONIC_2.0 - Dell POC: Upon reload, when the interface ip is same as virtual IP, we get "arp_update arping: libnet_get_ipaddr4(): ioctl(): Cannot assign requested address" error
Root Cause:
When the configuration is saved, the NEIGHBOR entry in the CONFIG_DB is
also saved. After reload the static arp entry is created in the ip
stack. Now when the arp_update script runs, it tries to refresh this arp
entry. Since the interface now has no ipv4 address the arping fails and
error messages reported in the bug are seen.

UT Result:
Verified on Belgrade-6 with physical and Vlan interfaces

Change-Id: I85d878cbd5ef1603ad3c61129764ec012be020fe