[JIRA SONIC-6560] Don't force-remove Docker images during cleanup phase
Removed the -f flag from the 'docker rmi -f' commands that are
being issued as part of the docker-cleanup target which runs
at the end of each build. Speculating that force-removing
these Docker images may be causing intermediate images to
be removed as well, even ones that might be in use by another
build running on the same build server.

Change-Id: I41ceb49645cfc3285aba8468a38cd51be4908e72