[JIRA SONIC-6073] ST: Reboot after a crash prints OrchAgent crash log to console from previous boot
The SONIC-6073 is an issue about an orchagent termination message from earlier
failure before the reboot, inappropriately displayed on the console of current
session.

Although there is a timestamp in the message body, it's definitely a confusing
message, in the meantime, there is also a termination message available from
supervisor-proc-exit-listener, which is realtime and works for all the cases
of process terminations, not only SIGABRT (core dumped).

And thus, the best approach for this should be dropping the filter for
'core dumped' from rsyslogd config.

Signed-off-by: Dante (Kuo-Jung) Su <dante.su@broadcom.com>
Change-Id: I635a1f33444c6b272224c83bb495f64a2b24e510