[JIRA SONIC-8038] Incorporate upstream rework of Docker version check
Docker CE version 19.xx and later changed the 'docker info' output
format slightly, causing the existing shell commands to no longer
work. This was found and fixed in upstream master, and is being
incorporated in the Broadcom SONiC branches to future-proof in
the event of a Docker version server upgrade.

Change-Id: I743952904b60e75cacc91717b35211ce1762876e