[JIRA SONIC-8350] Enable accepting router advertisements when performing ZTP over oob
When using IPv6 address assignment via DHCP, it is import to receive router
advertisements to build route table. Without that even if IPv6 address is
assigned, host will not be able to ping to other nodes on the network.

SONiC disables RA's on eth0 as default configuration. This change, enables
accepting RA's on eth0 when ZTP is operational.

Signed-off-by: Rajendra Dendukuri <rajendra.dendukuri@broadcom.com>
Change-Id: I04315a5bc8e728e452939c21eaac2daf6e8ee4e6