[JIRA SONIC-9208] cleanup BCM specific configs or defaults from Arlo build
When we build a sonic Image, the script build_debian install several Debian
packages using the apt-get utility. The file /etc/resolv.conf is then modified.
It does not make sense to keep this file, since it might contains some
sensitive information (dns information regarding the network on which the
image was built), and anyway this file is being created by the dhcp client
on bootup. The script is then taking care of deleting this file.

Change-Id: Ia1403e8db8aa7fbb90f7574cfaddf27d739dcccd