[JIRA SONIC-10178] Add branding file to specify product name, version, nos name
Support to allow build time option to specify
- Product Name: Optional SONiC product identifier for the image we wish to build
- NOS Name: Optional SONiC OS name for the image we wish to build
- Build version: Optional SONiC build version string we wish to use for the image

All these parameters can be specified in branding.mk file or passed on along with
the make command as an argument. The command line passed arguments take precedence
over the contents of branding.mk file.

Added utility build_pkg.sh which can be used to create an installer from an
existing SONiC installer binary and additional files. This can be used to include
user provided sonic_branding.yml. The sonic_branding.yml provided by user as part
of the package replaces the file in /etc/sonic/sonic_branding.yml which is created
at compile time.

The product name is included in the sonic_branding.yml as a value of the field
product_name.

The ONIE install script is enhanced to:
- Process user provided sonic_branding.yml
- Use NOS Name provided at build time or in sonic_branding.yml as part of pkg file
- Process SONiC pkg file

Change-Id: I1ad3e53b60f96cb1e3886910f227054d36ab0dc8