Allow REST server to listen on Unix socket
This change adds a -clicacert argument to the REST server. This flag
specifies a CA certificate against which the user certificates will be
signed, and if specified, the REST server will use this to verify the
client certificates. This is intended for use by the CLI client to
communicate with the REST server.