[JIRA SONIC-12047] Support of non-supported platform/profile of VxLAN
Enhanced SAI to indicate whether the corresponding platform/profile is supported for VxLAN feature.
Intorduced a new table in stateDb (VxlanCapability).
Orchagent will populate the VxlanCapability table to indicate enable/disable based on the SAI
indication.
Vxlanmgrd will read this table (while porcessing the conifugration). It will perform the following:
If the capability is unknown (orchagent did not get a chance to write into stateDb), then it will
not process the incoming config request. It will not flush the request from the sync queue and hence
the config will be retried again.
If the capability is disabled, it will ignore the incoming config request. Additioanlly, it will
also flush the request from the sync queue.
If the capability is enabled, it will process the incoming config request. It will flush the request
from the sync queue (if config process was successful).
Enhanced the logging/debugsh command to display the vxlanCapability to display
unknown/disable/enable for the corresponding platform/profile

Additional Fix:
With the previous checkin, it was experienced that when ztp is enabled and though it is in l3
profile, vxlan was disabled.
Made changes to config.bcm and the script to have the following new behavior:
L3 profile - vxlan is enabled
L2 profile - vxlan is disabled
Irrespective of the ztp enable/disable, the vxlan is enabled/disabled based on whether the profile
is l3/l3.

Test Cases:
End to end Vxlan with BUM/Unicast with multiple tunnels, multiple underlay next-hop, trigger
underlay link flaps, route flaps, clear ip bgp* and ensured traffic is restored for all 1000
vlans and 6000 macs.
Tested all combination of netinstall, software upgrade, changing of profile from l2 to l3 and
vice-verca and ensured that vxlan is enabled for l3 profile and disabled for the l2 profile.
The vxlan config is always accepted. However, vxlanmgrd ignores the config for non supported
platfrom/profile. It will process the request only for the supported platform/profile.

Change-Id: I1768f90f74e154370fe65ac8c9458ef124338d05