[JIRA SONIC-9668] fixed iccpd not to set port isolate when peer mclag interface is not up
fixed the following in iccpd code
1) when mclag interface is configured, port isolation was set without
considering whether peer interface up or not. fixed it to set only local and
peer mclag interface  is up
2) in standby node, when mclag interface is configured, if peer mclag interface
is configured, change the mac of mclag interface to be the peer system mac.
when mclag interface is removed on active or when peer mclag interface is
removed or when peer is disconnected mac will be reverted to standby's system
mac

verified on local setup

Change-Id: I94553edb59e047d7d75b4247a533a545c02ec030