[JIRA SONIC-14842] Increase systemd service start timeout for S6000
It is observed that the 120s value for DefaultTimeoutStartSec is not sufficient
for snmp/teamd systemd service to start. This is observed on S6000 switch
which has a slower CPU and services need more time to start. Similar change
was made to increase the timeout from 90s to 120s in SONIC-2.0.

Change-Id: Ibc729d0301c4e34aefa19405521004bc17a6e862