[JIRA SONIC-11271] Add fsck filesystem check support prior mounting filesystem
If the filesystem become non clean ("dirty"), SONiC does not run fsck to
repair and mark it as clean again.

This JIRA adds the functionality to run fsck on each boot, prior to the
filesystem being mounted. This allows the filesystem to be repaired if
needed.

Note that if the filesystem is marked as clean, fsck does nothing and simply
return.

How to verify this JIRA:

Using an image *without* this patch:

- Make the image non clean
(we are assuming that the platform uses /dev/sda3, please adjust depending of the platform)
dd if=/dev/sda3 of=superblock bs=1 count=2048
printf "$(printf '\\x%02X' 2)" | dd of="superblock" bs=1 seek=1082 count=1 conv=notrunc &> /dev/null
dd of=/dev/sda3 if=superblock bs=1 count=2048
- Verify the filesystem is dirty
tune2fs -l /dev/sda3 | grep "Filesystem state:"
- reboot
- verify that the filesystem is still dirty

Using the image with this patch:

- same test as above
after the next reboot, you should see the image being clean again

Change-Id: I9f4e55069b0c2502412f6f2194b6fe8528728aad