[JIRA SONIC-6201] SONiC configuration erase feature
This feature can be used to remove configuration files on a SONiC switch and bring it back to the default configuration state.
The following configuration files can be removed using this feature:

-Startup Configuration JSON - /etc/sonic/config_db.json
-Persistent application configuration files in the /etc/sonic directory like those used by FRR
-User installed debian packages including changes to any files in SONiC image on disk

After executing the  erase command, the user must reboot the switch for the changes to take effect.
The reboot command displays a warning message to the user indicating
the changes that will happen up on next reboot.

config erase - This command deletes the startup configuration JSON file
and all application configuration files in the
/etc/sonic directory. The management interface configuration
in the startup configuration file is retained so that the user
can access the switch using the same management address after
the switch reboot.

config erase boot - This command deletes the startup configuration
JSON file and all application configuration files in the /etc/sonic
directory. The management interface configuration in the startup
configuration JSON file is also removed. The SONiC switch boots
with a factory default configuration file.

config erase install - This command removes all changes made by the
user. All user installed packages and file changes are removed. It
also deletes the startup configuration JSON file and the files in
/etc/sonic directory. The SONiC switch is reverted to the same state
as a newly installed image. After the SONiC switch is rebooted,
if the Zero Touch Provisioning (ZTP) feature is enabled, the SONIC
switch will start performing ZTP to discover and download the switch
configuration.

config erase cancel - For the config erase command operation to take
effect, the user has to reboot the switch after issuing the config erase
command. If the user wishes to not proceed with configuration removal
operation, the write erase cancel command can be used to undo the
previously issued write erase command.

Change-Id: I2d1755f59fc530b50eab3719f71e00beffcba6e8