[SONIC-14505] Disallow launching CLI for root user
Prior to the introduction of RBAC and certificate authentication for CLI
users, all users were treated as "admin". However, after the
introduction of certificate authentication, only users created by the
system have the certificates set up for letting the CLI session
authenticate with the REST server.

These certificates are not created for the root user, which means that
if the root launches clish, there are no certificates available, and
then running any command would cause the REST server to reject the
connection as unauthenticated.

This change checks if the current user is root, and if so, aborts
launching the CLI with an error message.