[SONIC-14938] Fix correct shell for management users
Prior to this change, all users were getting dropped into the bash
shell. However, this is a security hole, since operator users are only
allowed to read the state, and having access to the bash shell will
allow them to also modify/write the configuration.

While we could change the default shell on a per-user basis based on
their primary role, this creates an issue for remote users, since the
primary role is not known until the time the shell gets launched. This
would lead us to a situation where a user's privilege level gets
modified on the RADIUS/TACACS+ server, but the default shell is not
changed locally.

This change instead drops every user into a script that launches the
real shell based on the user's primary group, i.e., users with a default
primary group of `admin` would get dropped into bash, while `operator`
users would get dropped into sonic-cli. As a precaution against invalid
primary groups, users with a primary group other than `admin` or
`operator` are not allowed to login - the default shell is `/bin/false`.