[JIRA SONIC-15464] Fixed ICCP peer socket snd and receive buffer sizes
increased peer link socket message buffer sizes to avoid peer socket send
issues.

Root cause details:

between iccp peer there are lot of message exchange and if the peer node is busy
not processing the message for a long time. sender end is not to send the
packets to the peer since the tcp stream socket buffer is not available and it
sends partial msg to the peer. if the peer receives partial message with header
info. it can use header to look for msg_len of the tlv data. but it would endup
stepping on the other messages in queue and it corrupts the sys pointer.

currently increase the send and recv buffer sizes to 2M byes  to handle this
condition.but to properly fix this scenario plan to retry sending end.

tried with 100k macs. tried flapping peer link interface

Change-Id: I2f34e43f2c513cbabdf5209c1fc423e8d3bad6bc