[JIRA SONIC-17372] KLISH CLI API documentation for kdump, config erase and ZTP
KLISH CLI documentation is generated automatically from the XML file.

We are updating these XML "description" statements to have the documentation
generated for both kdump, config erase and ZTP.

In addition to this, we are fixing a small defect (not yet reported):

in write-erase command, when a confirmation [y/N] is asked, we were only
waiting specifically for "y" or "N". Both "Y" or "n" should be also accepted.

Change-Id: Icb35f06f3d08c571c02a0cfa07a056a98cf6e8e5