Relax restrictions on CLI certificate (#589)
Prior to this change, a reset of the system clock would potentially
cause the CLI client to fail to connect to the REST server, since the
system clock would typically reset to a time earlier than the validity
period of the user certificate. This would result in the TLS handshake
failing since the client certificate would be rejected.

This change relaxes the certificate checks for the client certificate
and only requires that the CLI client send a certificate. While this
does make the system slightly less secure, it is only applicable to the
listener for the CLI interface. Therefore, this does not expose any new
security holes for the REST server.